Enum evm_gasometer::GasCost

pub enum GasCost {

    Zero,
    Base,
    VeryLow,
    Low,
    Invalid,
    ExtCodeSize {
        target_is_cold: bool,
    },
    Balance {
        target_is_cold: bool,
    },
    BlockHash,
    ExtCodeHash {
        target_is_cold: bool,
    },
    Call {
        value: U256,
        gas: U256,
        target_is_cold: bool,
        target_exists: bool,
    },
    CallCode {
        value: U256,
        gas: U256,
        target_is_cold: bool,
        target_exists: bool,
    },
    DelegateCall {
        gas: U256,
        target_is_cold: bool,
        target_exists: bool,
    },
    StaticCall {
        gas: U256,
        target_is_cold: bool,
        target_exists: bool,
    },
    Suicide {
        value: U256,
        target_is_cold: bool,
        target_exists: bool,
        already_removed: bool,
    },
    SStore {
        original: H256,
        current: H256,
        new: H256,
        target_is_cold: bool,
    },
    Sha3 {
        len: U256,
    },
    Log {
        n: u8,
        len: U256,
    },
    ExtCodeCopy {
        target_is_cold: bool,
        len: U256,
    },
    VeryLowCopy {
        len: U256,
    },
    Exp {
        power: U256,
    },
    Create,
    Create2 {
        len: U256,
    },
    SLoad {
        target_is_cold: bool,
    },

}

Gas cost.

Variants

Zero

Zero gas cost.

Base

Base gas cost.

VeryLow

Very low gas cost.

Low

Low gas cost.

Invalid

Fail the gasometer.

ExtCodeSize

Fields

target_is_cold: bool

True if address has not been previously accessed in this transaction

Gas cost for EXTCODESIZE.

Balance

Fields

target_is_cold: bool

True if address has not been previously accessed in this transaction

Gas cost for BALANCE.

BlockHash

Gas cost for BLOCKHASH.

ExtCodeHash

Fields

target_is_cold: bool

True if address has not been previously accessed in this transaction

Gas cost for EXTBLOCKHASH.

Call

Fields

value: U256

Call value.

gas: U256

Call gas.

target_is_cold: bool

True if target has not been previously accessed in this transaction

target_exists: bool

Whether the target exists.

Gas cost for CALL.

CallCode

Fields

value: U256

Call value.

gas: U256

Call gas.

target_is_cold: bool

True if target has not been previously accessed in this transaction

target_exists: bool

Whether the target exists.

Gas cost for `CALLCODE.

DelegateCall

Fields

gas: U256

Call gas.

target_is_cold: bool

True if target has not been previously accessed in this transaction

target_exists: bool

Whether the target exists.

Gas cost for DELEGATECALL.

StaticCall

Fields

gas: U256

Call gas.

target_is_cold: bool

True if target has not been previously accessed in this transaction

target_exists: bool

Whether the target exists.

Gas cost for STATICCALL.

Suicide

Fields

value: U256

Value.

target_is_cold: bool

True if target has not been previously accessed in this transaction

target_exists: bool

Whether the target exists.

already_removed: bool

Whether the target has already been removed.

Gas cost for SUICIDE.

SStore

Fields

original: H256

Original value.

current: H256

Current value.

new: H256

New value.

target_is_cold: bool

True if target has not been previously accessed in this transaction

Gas cost for SSTORE.

Sha3

Fields

len: U256

Length of the data.

Gas cost for SHA3.

Log

Fields

n: u8

Topic length.

len: U256

Data length.

Gas cost for LOG.

ExtCodeCopy

Fields

target_is_cold: bool

True if target has not been previously accessed in this transaction

len: U256

Length.

Gas cost for EXTCODECOPY.

VeryLowCopy

Fields

len: U256

Length.

Gas cost for some copy opcodes that is documented as VERYLOW.

Exp

Fields

power: U256

Power of EXP.

Gas cost for EXP.

Create

Gas cost for CREATE.

Create2

Fields

len: U256

Length.

Gas cost for CREATE2.

SLoad

Fields

target_is_cold: bool

True if target has not been previously accessed in this transaction

Gas cost for SLOAD.